refactor: 상수화, 패키지 정리, handler의 interface화
