+++ src/main/java/annotations/APIMapping.java
@@ -1,6 +1,6 @@
 package annotations;
 
-import webserver.HttpMethod;
+import http.HttpMethod;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface HandleRequest {
+public @interface APIMapping {
     String path();
     HttpMethod httpMethod();
 }
