+++ src/main/java/constant/Constant.java
@@ -1,0 +1,13 @@
+package constant;
+
+public class Constant {
+	public static final String QUERY_SEPARATOR = "?";
+	public static final String PARAMETER_SEPARATOR = "&";
+	public static final String SPACE = " ";
+	public static final String HEADER_SEPARATOR = ": ";
+	public static final String PARAMETER_EQUAL_SIGN = "=";
+	public static final String EXTENSION_SEPARATOR = ".";
+	public static final String HEADER_CONTENT_LENGTH = "Content-Length";
+	public static final String HEADER_CONTENT_TYPE = "Content-Type";
+	public static final String CONTENT_TYPE_FORM = "application/x-www-form-urlencoded";
+}
