+++ src/main/java/hadler/Handler.java
@@ -1,0 +1,8 @@
+package hadler;
+
+import http.HttpRequest;
+import http.HttpResponse;
+
+public interface Handler {
+	HttpResponse handle(HttpRequest httpRequest);
+}
