+++ src/main/java/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 public enum HttpMethod {
     GET, POST,
