+++ src/main/java/http/StatusCode.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 public enum StatusCode {
     OK(200, "OK"),
