+++ src/main/java/utils/FormUrlEncodedParsingStrategy.java
@@ -1,5 +1,7 @@
 package utils;
 
+import static constant.Constant.*;
+
 import java.util.Arrays;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -13,8 +15,8 @@ public class FormUrlEncodedParsingStrategy implements HttpBodyParsingStrategy {
 
         Map<String, String> map;
         try {
-            map = Arrays.stream(body.split("&"))
-                    .map(s -> s.split("="))
+            map = Arrays.stream(body.split(PARAMETER_SEPARATOR))
+                    .map(s -> s.split(PARAMETER_EQUAL_SIGN))
                     .map(arr -> Map.entry(arr[0], arr[1]))
                     .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
         } catch (Exception e) {
